UI Debugging and Changes
Reversed the picker view so that the user pulls their finger down to navigate the picker (instead of up). This makes it so (theoretically) they don't accidentally get the control center to appear most of the time. Fixed an issue where clearing a text field would cause that text field to start editing.